requirements.txt
argcomplete==2.0.0
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
blinker==1.4
boto==2.49.0
boto3==1.21.15
botocore==1.24.15
certifi==2021.10.8
cfn-flip==1.3.0
charset-normalizer==2.0.12
click==8.0.4
Django==4.0.3
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-dynamodb-sessions==0.6
django-s3-storage==0.13.6
djangorestframework==3.13.1
durationpy==0.5
dynamorm==0.11.0
future==0.18.2
hjson==3.0.2
html5lib==1.1
idna==3.3
importlib-resources==5.4.0
jmespath==0.10.0
jsonschema==4.4.0
kappa==0.6.0
lxml==4.8.0
marshmallow==3.14.1
placebo==0.9.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2021.3
PyYAML==6.0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.63.0
troposphere==3.2.2
urllib3==1.26.8
webencodings==0.5.1
Werkzeug==2.0.3
wsgi-request-logger==0.4.6
zappa==0.54.1
zipp==3.7.0
